ditch nulls, fix NPE
